+++ packages/shared/analytics/hook-observable.ts
@@ -1,29 +1,10 @@
-import { DEBUG_ANALYTICS } from 'config'
 import { Vector2 } from '@dcl/ecs-math'
 
-import { worldToGrid } from 'atomicHelpers/parcelScenePositions'
-import { defaultLogger } from 'shared/logger'
-
-import { avatarMessageObservable } from './comms/peers'
-import { AvatarMessageType } from './comms/interface/types'
-import { positionObservable } from './world/positionThings'
-import { globalObservable } from './observables'
-
-export type SegmentEvent = {
-  name: string
-  data: string
-}
-
-export function trackEvent(eventName: string, eventData: Record<string, any>) {
-  if (DEBUG_ANALYTICS) {
-    defaultLogger.info(`Tracking event "${eventName}": `, eventData)
-  }
-
-  globalObservable.emit('trackingEvent', {
-    eventName,
-    eventData
-  })
-}
+import { worldToGrid } from '../../atomicHelpers/parcelScenePositions'
+import { avatarMessageObservable } from '../comms/peers'
+import { AvatarMessageType } from '../comms/interface/types'
+import { positionObservable } from '../world/positionThings'
+import { trackEvent } from '.'
 
 const TRACEABLE_AVATAR_EVENTS = [
   AvatarMessageType.ADD_FRIEND,
